[staging] CassandraSink - Write job results to Cassandra
The CassandraSink class allows writing Tuples to a Cassandra cluster.

It is assumed the destination keyspace/table already exist in the Cassandra cluster.

It is possible to specify which tuples map to which column in the Cassandra table,
and which item in the tuple is to be used as Cassandra's Partition key.

CassandraSink is aware of the cluster topology, and partitions the data accordingly.
This results in efficient data transfers, because the data is sent only once and directly
to the node that owns it.