[hotfix][runtime] Cherry-picked a21af67: --- A customized filename can be specified through Content-Disposition that also allows passing of path information which was not properly handled. This is fixed now.
We just use the filename instead of interpreting any path information that was passed through a custom filename. Two tests were added to verify the proper behavior:
1. a custom filename without path information was used
2. a custom filename with path information was used

The change required adapting the MultipartUploadResource in a way that it is used not as a @ClassRule but as a @Rule instead. This enables us to initialize it differently on a per-test level. The change makes the verification of the uploaded files configurable.